requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
av==14.4.0
beautiful-date==2.3.0
beautifulsoup4==4.13.4
bert-score==0.3.13
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
gcsa==2.5.1
geocoder==1.38.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
googlemaps==4.10.0
googlesearch-python==1.3.0
gradio==5.30.0
gradio_client==1.10.1
greenlet==3.2.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.2
hk-bus-eta==2.1.6
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
imutils==0.5.4
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langsmith==0.3.43
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
ollama==0.5.1
omegaconf==2.3.0
openai==1.82.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.5
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-weather==2.1.0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
pyzmq==27.0.2
qwen-vl-utils==0.0.11
ratelim==0.1.6
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge-chinese==1.0.3
rouge_score==0.1.2
rsa==4.9.1
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.6.0+cu124
torchvision==0.22.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.1
triton==3.3.0
trl==0.9.6
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2025.2
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wikipedia==1.4.0
xformers==0.0.29.post1
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0
